+++ api/blueprints/__init__.py
@@ -2,7 +2,7 @@ from .drbCitation import citation
 from .drbCollection import collection
 from .drbEdition import edition
 from .drbInfo import info
-from .links import link
+from .drbLink import link
 from .drbOPDS2 import opds
 from .drbSearch import search
 from .drbUtils import utils
